Drop use of readlink for a custom function
The readlink command is non-standard and varies from unix to unix, if it
exists at all. Most are pretty reasonable, except for BusyBox, which
ships with a broken version of readlink (prompting this commit). Unlike
the others, BusyBox readlink returns an error if the path doesn't
currently exist, which makes it useless for qpkg.

The new readlpath function here isn't nearly as fancy but it's
sufficient for qpkg.