bashrc: Set ASAN_OPTIONS to better defaults
It's crazy that it would do anything *but* abort on error, since the
default behavior is useless in debuggers. I want the program to trap so
that I can look around at the still-living process. I would also prefer
it print even less information, but I'll have to make due with just
omiting the legend.