vimrc: Use w64devkit shell when available
All this time I haven't been using sh as 'shell' when running Vim under
w64devkit. I had worked out the configuration, but never bothered to put
it in my vimrc until now. My recent unconditional change to makeprg was
incompatible with cmd.exe, reminding me to finally address it.

Note that Vim is still pretty silly about all this: It invokes sh by way
of cmd.exe — it's the only way it knows how to start processes — and so
internally it must compute an extra layer of quoting for all commands.

Ref: https://nullprogram.com/blog/2022/02/18/
Ref: https://old.reddit.com/r/vim/comments/r6dthv/_/hmtj93p/