vim: Set debug-oriented config for w64devkit
I always want a debug build when building from Vim, so set CFLAGS and
LDFLAGS accordingly and override the makefile (-e). In the single source
file case I can skip the makefile altogether and instead mp+=%:r to rely
on GNU Make's built-in rules to do all the work. If there is an actual
makefile, it will be configured for a release by default, and releases
are always built outside of Vim.

This configuration has great ergonomics combined with GDB: fast builds,
implicit configuration, a good warning set, traps directly on undefined
behavior with no excess stack frames.