gdb: Strip implementation stack frames on abort
ASan irritates me by placing 8 or so frames on top of the actual defect,
and even glibc assert() leaves 3 junk frames in the way. This makes for
a poor UI experience as I have to dig through these frames each time an
assertion is triggered. Why is this the default behavior?! This hook
strips away the implementation's frames the same way I've done literally
thousands of times by hand.

This has worked well so far, making GDB more pleasant. Unfortunately it
requires Python ($_caller_matches), which generally means it won't work
on Windows.