gdbinit: Use _any_caller_matches and _thread
The $_caller_matches convenience function does not match the current
frame, but $_any_caller_matches does match the current frame. This is
probably a GDB bug, but it's a useful one. It's also a little more
robust.

Unfortunately when a process exits, hook-stop gets a Python exception
from the frame matcher since there are no frames. This is bad design
given that it's impossible to handle errors in scripts. It should simply
not match. As a workaround test $_thread, which will be zero when the
process is not running.