bash: Configure UBSan to abort as well
I couldn't get this to work for the longest time, but then I noticed afl
sets both abort_on_error and halt_on_error. Both are required to abort
on error, which I suspect is a UBSan bug.

Also drop MSAN_OPTIONS while I'm at it. MSan continues to disappoint,
and I've given up on trying to get something useful out of it.